+++ gridopt/power_flow/__init__.py
@@ -6,28 +6,25 @@
 # GRIDOPT is released under the BSD 2-clause license. #
 #*****************************************************#
 
-from .nr_pf import NRPF
 from .dc_pf import DCPF
 from .dc_opf import DCOPF
-from .augl_pf import AugLPF
-from .augl_opf import AugLOPF
-from .ipopt_opf import IpoptOPF
+from .ac_pf import ACPF
+from .ac_opf import ACOPF
 from .method import PFmethod
 from .method_error import PFmethodError
 
-methods = [NRPF,DCPF,DCOPF,AugLPF,AugLOPF,IpoptOPF]
+methods = [DCPF,DCOPF,ACPF,ACOPF]
 
 def new_method(name):
     """
-    Creates a power flow or optimal power flow
-    method.
+    Creates a power flow or optimal power flow method.
     
     Parameters
     ----------
-    name : string
+    name : {'DCPF','DCOPF','ACPF','ACOPF'}
     """
     
     try:
         return methods[list([x.name for x in methods]).index(name)]()
     except ValueError:
-        raise ValueError('invalid PF method name')
+        raise ValueError('invalid method name')