+++ CHANGELOG.md
@@ -4,6 +4,14 @@ Targets
 Unreleased
 ----------
 
+Version 1.3.3
+-------------
+* Unified methods (ACPF, ACOPF, DCPF, DCOPF) with param "optsolver" for choosing optimization solver.
+* Added documentation for "gridopt" command-line utility.
+* Updated and re-enabled DCOPF unittests.
+* Added new OPTALG solver "inlp" to ACPF and ACOPF and updated documentation.
+* limit_gens flag for non-NR ACPF with unittest.
+
 Version 1.3.2
 -------------
 * Compatibility with PFNET v1.3.0.